code/requirements.txt
django
djangorestframework
pandas
numpy
scikit-learn
transformers
torch
openpyxl
openai
langchain
